interfaceAuto/atf/requirements.txt
BeautifulReport==0.1.3
requests==2.22.0
xlrd==1.2.0
pyyaml==5.4.1
openpyxl
apscheduler
nose-parameterized==0.6.0
beautifulsoup4==4.10.0